Fixed World Editor build on Windows in VS2012.
Visual Studio 2012 doesn't support initializer lists for complex data types.
It also is nitpicky about other syntax things.

Also disabled some warnings that were being treated as errors via the Bullet
headers.

+++ src/editor_controller.cpp
@@ -32,8 +32,9 @@ void EditorController::Update() {
     vec2 delta = editor_config_->mouse_sensitivity() *
                  vec2(input_system_->get_pointers()[0].mousedelta);
 
-    vec3 side_axis =
-        quat::FromAngleAxis(-M_PI / 2, mathfu::kAxisZ3f) * facing_current_;
+    vec3 side_axis = quat::FromAngleAxis(-static_cast<float>(M_PI) / 2.0f,
+                                         mathfu::kAxisZ3f) *
+                     facing_current_;
 
     quat pitch_adjustment = quat::FromAngleAxis(-delta.y(), side_axis);
     quat yaw_adjustment = quat::FromAngleAxis(-delta.x(), mathfu::kAxisZ3f);
